requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.23.2
pydantic==1.10.12
PyJwt==1.7.1
python-decouple==3.3